refactor money

+++ src/lib/transaction/transaction.utils.ts
@@ -4,7 +4,7 @@ export function buildTransaction(formData: FormData): Transaction {
 	return {
 		id: Number(formData.get('id')),
 		description: formData.get('description') as string,
-		amount: parseFloat(formData.get('amount') as string),
+		amount: parseInt(formData.get('amount') as string),
 		date: formData.get('date') as string,
 		isPaid: formData.get('isPaid') === 'on',
 		referenceDate: formData.get('referenceDate') as string,
